Clean up config dependencies, eagerly load the batch job registerer
