+++ plugin/cr/src/main/java/org/opencds/cqf/ruler/cr/CrConfig.java
@@ -5,17 +5,17 @@ import org.opencds.cqf.cql.evaluator.measure.MeasureEvaluationOptions;
 import org.opencds.cqf.external.annotations.OnDSTU3Condition;
 import org.opencds.cqf.external.annotations.OnR4Condition;
 import org.opencds.cqf.ruler.cql.CqlConfig;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 import ca.uhn.fhir.context.FhirContext;
 
 @Configuration
-@ConditionalOnBean(CqlConfig.class)
 @ConditionalOnProperty(prefix = "hapi.fhir.cr", name = "enabled", havingValue = "true", matchIfMissing = true)
+@Import(CqlConfig.class)
 public class CrConfig {
 	@Bean
 	public CrProperties crProperties() {
