+++ plugin/cr/src/test/java/org/opencds/cqf/ruler/cr/r4/provider/CareGapsProviderIT.java
@@ -9,6 +9,7 @@ import org.hl7.fhir.r4.model.Parameters;
 import org.hl7.fhir.r4.model.StringType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.opencds.cqf.ruler.cql.CqlConfig;
 import org.opencds.cqf.ruler.cr.CrConfig;
 import org.opencds.cqf.ruler.test.RestIntegrationTest;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -16,7 +17,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import ca.uhn.fhir.rest.server.exceptions.InternalErrorException;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = { CareGapsProviderIT.class,
-		CrConfig.class }, properties = {
+		CrConfig.class, CqlConfig.class }, properties = {
 				"hapi.fhir.fhir_version=r4", "hapi.fhir.enforce_referential_integrity_on_write=false",
 				"hapi.fhir.enforce_referential_integrity_on_delete=false", "hapi.fhir.cr.enabled=true",
 				"hapi.fhir.cr.measure_report.care_gaps_reporter=Organization/alphora",
@@ -93,12 +94,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-				getClient().operation().onType(Measure.class).named("$care-gaps")
-						.withParameters(params)
-						.useHttpGet()
-						.returnResourceType(Parameters.class)
-						.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -111,12 +111,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-				getClient().operation().onType(Measure.class).named("$care-gaps")
-						.withParameters(params)
-						.useHttpGet()
-						.returnResourceType(Parameters.class)
-						.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -128,12 +127,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-				getClient().operation().onType(Measure.class).named("$care-gaps")
-						.withParameters(params)
-						.useHttpGet()
-						.returnResourceType(Parameters.class)
-						.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -146,12 +144,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-				getClient().operation().onType(Measure.class).named("$care-gaps")
-						.withParameters(params)
-						.useHttpGet()
-						.returnResourceType(Parameters.class)
-						.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -164,12 +161,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@Test
@@ -202,12 +198,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -220,12 +215,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -239,12 +233,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 		params.addParameter().setName("practitioner").setValue(new StringType(practitionerValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -258,12 +251,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 		params.addParameter().setName("organization").setValue(new StringType(organizationValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@Test
@@ -276,12 +268,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("organization").setValue(new StringType(organizationValid));
 		params.addParameter().setName("practitioner").setValue(new StringType(practitionerValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 		// TODO: implement practitioner and organization
 		// assertDoesNotThrow(() -> {
 		// getClient().operation().onType(Measure.class).named("$care-gaps")
@@ -301,12 +292,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 		params.addParameter().setName("organization").setValue(new StringType(organizationValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 		// TODO: implement organization
 		// assertDoesNotThrow(() -> {
 		// getClient().operation().onType(Measure.class).named("$care-gaps")
@@ -327,12 +317,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 		params.addParameter().setName("practitioner").setValue(new StringType(practitionerValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -346,12 +335,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -363,12 +351,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("subject").setValue(new StringType(subjectPatientValid));
 		params.addParameter().setName("status").setValue(new StringType(statusValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -381,12 +368,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("status").setValue(new StringType(statusInvalid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@SuppressWarnings("java:S5778")
@@ -398,12 +384,11 @@ class CareGapsProviderIT extends RestIntegrationTest {
 		params.addParameter().setName("subject").setValue(new StringType(subjectPatientValid));
 		params.addParameter().setName("measureId").setValue(new StringType(measureIdValid));
 
-		assertThrows(InternalErrorException.class, () ->
-			getClient().operation().onType(Measure.class).named("$care-gaps")
-					.withParameters(params)
-					.useHttpGet()
-					.returnResourceType(Parameters.class)
-					.execute());
+		assertThrows(InternalErrorException.class, () -> getClient().operation().onType(Measure.class).named("$care-gaps")
+				.withParameters(params)
+				.useHttpGet()
+				.returnResourceType(Parameters.class)
+				.execute());
 	}
 
 	@Test
