+++ plugin/cr/src/test/java/org/opencds/cqf/ruler/cr/r4/provider/CollectDataProviderIT.java
@@ -14,6 +14,7 @@ import org.hl7.fhir.r4.model.Patient;
 import org.hl7.fhir.r4.model.Reference;
 import org.junit.jupiter.api.Test;
 import org.opencds.cqf.ruler.cr.CqlBuilder;
+import org.opencds.cqf.ruler.cr.CrConfig;
 import org.opencds.cqf.ruler.cr.r4.Libraries;
 import org.opencds.cqf.ruler.cr.r4.MeasureBuilder;
 import org.opencds.cqf.ruler.cr.r4.Patients;
@@ -24,7 +25,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 import ca.uhn.fhir.jpa.partition.SystemRequestDetails;
 
-@SpringBootTest(classes = { CollectDataProviderIT.class }, properties = { "hapi.fhir.fhir_version=r4", })
+@SpringBootTest(classes = { CollectDataProviderIT.class, CrConfig.class }, properties = {
+		"hapi.fhir.fhir_version=r4", })
 class CollectDataProviderIT extends DaoIntegrationTest {
 
 	@Autowired
@@ -53,22 +55,24 @@ class CollectDataProviderIT extends DaoIntegrationTest {
 		this.create(john);
 
 		Observation obs = newResource(Observation.class)
-			.setSubject(new Reference(john));
+				.setSubject(new Reference(john));
 		this.create(obs);
 
 		Encounter enc = newResource(Encounter.class)
-			.setSubject(new Reference(john));
+				.setSubject(new Reference(john));
 		this.create(enc);
 
 		// Submit it
 		Parameters results = collectDataProvider.collectData(new SystemRequestDetails(), m.getIdElement(), "2019-01-01",
 				"2019-12-31", Ids.simple(john), null, null);
 
-		List<ParametersParameterComponent> resources = org.opencds.cqf.ruler.utility.r4.Parameters.getPartsByName(results, "resource");
+		List<ParametersParameterComponent> resources = org.opencds.cqf.ruler.utility.r4.Parameters.getPartsByName(results,
+				"resource");
 		assertEquals(1, resources.size());
 		assertEquals("Observation", resources.get(0).getResource().fhirType());
 
-		List<ParametersParameterComponent> reports = org.opencds.cqf.ruler.utility.r4.Parameters.getPartsByName(results, "measureReport");
+		List<ParametersParameterComponent> reports = org.opencds.cqf.ruler.utility.r4.Parameters.getPartsByName(results,
+				"measureReport");
 		assertEquals(1, reports.size());
 		assertEquals("MeasureReport", reports.get(0).getResource().fhirType());
 	}
