+++ plugin/ra/src/main/java/org/opencds/cqf/ruler/ra/RAConfig.java
@@ -2,6 +2,7 @@ package org.opencds.cqf.ruler.ra;
 
 import org.opencds.cqf.external.annotations.OnR4Condition;
 import org.opencds.cqf.ruler.api.OperationProvider;
+import org.opencds.cqf.ruler.cr.CrConfig;
 import org.opencds.cqf.ruler.ra.r4.AssistedServlet;
 import org.opencds.cqf.ruler.ra.r4.ResolveProvider;
 import org.opencds.cqf.ruler.ra.r4.RiskAdjustmentProvider;
@@ -11,9 +12,11 @@ import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 @Configuration
 @ConditionalOnProperty(prefix = "hapi.fhir.ra", name = "enabled", havingValue = "true", matchIfMissing = true)
+@Import(CrConfig.class)
 public class RAConfig {
 
 	@Bean
